pyproject.toml
[project]
name = "fraudpheus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.8",
    "httpx>=0.27.0",
    "pyairtable>=3.3.0",
    "python-dotenv>=1.2.1",
    "slack-bolt>=1.27.0",
    "slack-sdk>=3.40.0",
    "uvicorn>=0.40.0",
]
